+++ platformio/managers/core.py
@@ -21,9 +21,9 @@ from platformio import __version__, exception, util
 from platformio.managers.package import PackageManager
 
 CORE_PACKAGES = {
-    "contrib-piohome": ">=0.7.0,<2",
+    "contrib-piohome": ">=0.7.1,<2",
     "contrib-pysite": ">=0.1.5,<2",
-    "tool-pioplus": ">=0.14.1,<2",
+    "tool-pioplus": ">=0.14.3,<2",
     "tool-unity": "~1.20302.1",
     "tool-scons": "~2.20501.3"
 }