+++ HISTORY.rst
@@ -14,6 +14,9 @@ PlatformIO 3.0
 * Show all available upload protocols before firmware uploading in a build log
 * Handle "os.mbed.com" URL as a Mercurial (hg) repository
 * Improved support for old mbed libraries without manifest
+* Fixed project generator for Qt Creator IDE
+  (`issue #1303 <https://github.com/platformio/platformio-core/issues/1303>`_,
+   `issue #1323 <https://github.com/platformio/platformio-core/issues/1323>`_)
 * Fixed issue with duplicated "include" records when generating data for IDE
   (`issue #1301 <https://github.com/platformio/platformio-core/issues/1301>`_)