+++ HISTORY.rst
@@ -7,6 +7,7 @@ PlatformIO 3.0
 3.5.2 (2018-??-??)
 ~~~~~~~~~~~~~~~~~~
 
+* Multiple themes (Dark & Light) for `PlatformIO Home <http://docs.platformio.org/page/home/index.html>`__
 * Control `PIO Unified Debugger <http://docs.platformio.org/en/latest/plus/debugging.html>`__
   and its firmware loading mode using
   `debug_load_mode <http://docs.platformio.org/en/latest/projectconf/section_env_debug.html#debug-load-mode>`__ option